// src/common/types/enums.ts
export enum Role {
    CLIENT = "CLIENT",
    DEVELOPER = "DEVELOPER",
    ADMIN = "ADMIN",
}

export enum DevLevel {
    EXPERT = 'EXPERT',
    MID = 'MID',
    FRESHER = 'FRESHER',
}

export enum AssignmentStatus {
    PENDING = 'PENDING',
    ACCEPTED = 'ACCEPTED',
    REJECTED = 'REJECTED',
    EXPIRED = 'EXPIRED',
}

export enum VerificationStatus {
    PENDING = 'PENDING',
    APPROVED = 'APPROVED',
    REJECTED = 'REJECTED'
};


export enum PackageCode {
    FREE = 'FREE',
    BASIC = 'BASIC',
    STANDARD = 'STANDARD',
    PREMIUM = 'PREMIUM',
}

export enum SubscriptionStatus {
    ACTIVE = 'ACTIVE',
    CANCELED = 'CANCELED',
    EXPIRED = 'EXPIRED',
    PENDING = 'PENDING',
    SCHEDULED = 'SCHEDULED',
}